kafka/requirements.txt
kafka-python==2.0.2
psycopg2-binary
python-dotenv